# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "KARBONİZGİ",
        "contact": {}
    },
    "host": "https://api.karbonizgi.tr",
    "paths": {
...
        "/chat": {
            "post": {
                "description": "Chat Kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "User chat Kısmı",
                "parameters": [
                    {
                        "description": "Kullanıcının user id bilgisini ve ilk başta score bilgisini girmelisin.Score bilgisi /score endpointinde data kısmında dönüyor.Eğer kullanıcı konuşmayı devam ettirirse message kısmında kullanıcının mesajını gönderebilirsin. Filter kısmında main ya da detail ifadelerini göndermelisin. main için key kısmını girmene gerek yok ama detail kısmında key kısmına alt başlığın key bilgisini, score kısmına da alt başlıkta aldığı skoru göndermelisin.",
                        "name": "PersonInf",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.PersonInf"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message kısmını kullanıcıya gösterebilirsin.Eğer kullanıcı karbonayak izi dışında sorular sorarsa chat cevap vermez karbon ayak izine yönlendirir.2 kere gereksiz soru sorulursa data kısmında 'false' döner ve burada chat konuşmasını bitir sonra message kısmını kullanıcıya gösterebilirsin.",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
        "/comments": {
            "get": {
                "description": "Kullanıcı soru-cevap kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "person ya da company ifadesi girilir.",
                        "name": "user_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kullanıcı eleştiri soru cevap kısımları bulunur.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/controllers.Comment"
                            }
                        }
                    }
                }
            }
        },
...
        "/general-chat": {
            "post": {
                "description": "Chat Kısmı",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "User genel chat Kısmı",
                "parameters": [
                    {
                        "description": "kullanıcı mesajını girmelisin.",
                        "name": "GeneralReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controllers.GeneralReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message içerisinde yanıt döner",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/controllers.Response"
                        }
                    }
                }
            }
        },
...
        "/suggested": {
            "get": {
                "description": "Örnek aksiyonlar",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Data"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "person ya da company ifadesi girilir.",
                        "name": "user_type",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Kullanıcıya aksiyon önerileri için kullanılır.s",
                        "schema": {
                            "$ref": "#/definitions/controllers.SuggestResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controllers.Comment": {
            "type": "object",
            "properties": {
                "answer": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                }
            }
        },
...
        "controllers.GeneralReq": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "controllers.PersonInf": {
            "type": "object",
            "properties": {
                "filter": {
                    "type": "string"
                },
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "score": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "controllers.Response": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "controllers.SuggestResp": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}